requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-jose[cryptography]
pydantic>=2.0
